Many updates (oops)
* added strings
* renamed some methods
* added s-expressions
* added parsing
* changed interpreter imports for fancy names (may change)

Some features are not so good at the moment. S-Expressions all contain
other types... which as I commit this I doubt is the correct way to go
about things. It seems to me that perhaps I just need better s-exp->XYZ
conversion methods (and vice-versa?), and the parser should not just
throw everything into an s-expression and then interpret it.